requirements.txt
# Claude Code Launcher - Python Dependencies
# Updated after Flet system tray testing failure
# Framework: Tkinter + ttkbootstrap + pystray

# UI Framework
ttkbootstrap==1.10.1  # Modern themed Tkinter with Windows 11 style themes

# System Tray
pystray==0.19.5       # System tray icon support (Windows, macOS, Linux)
Pillow==11.0.0        # Image handling for tray icons

# Async Operations
aiohttp==3.11.0       # Async HTTP client for NPM registry validation

# Windows Integration
pywin32==308          # Windows API access (registry, credential manager)
keyring==25.5.0       # Secure credential storage using Windows Credential Manager

# Process Management
psutil==6.1.0         # Cross-platform process utilities (Claude Code detection)

# Data Handling
python-dateutil==2.9.0 # Date parsing and manipulation
jsonschema==4.23.0    # JSON schema validation for config files

# Testing
pytest==8.4.2         # Test framework
pytest-asyncio==0.25.2 # Async test support

# Note: Tkinter comes bundled with Python (no need to install separately)